Move Spotify hook logic in new authentication hook The Spotify hook will be used to make simple Spotify API calls, not to connects users Update authentication system
